+++ Pandora/Paradigm/Algebraic.hs
@@ -2,7 +2,6 @@
 module Pandora.Paradigm.Algebraic (module Exports, type (:+*+:), type (:*+*:), type (:++*:)) where
 
 import Pandora.Paradigm.Algebraic.Functor as Exports
-import Pandora.Paradigm.Algebraic.Exponential as Exports
 import Pandora.Paradigm.Algebraic.Product as Exports
 import Pandora.Paradigm.Algebraic.Sum as Exports
 
@@ -17,6 +16,7 @@ import Pandora.Pattern.Functor.Semimonoidal (Semimonoidal (mult))
 import Pandora.Pattern.Functor.Monoidal (Monoidal (unit))
 import Pandora.Pattern.Functor.Comonad (Comonad)
 import Pandora.Pattern.Functor.Traversable (Traversable ((<-/-)))
+import Pandora.Pattern.Operation.Exponential (type (-->), type (--<))
 import Pandora.Pattern.Operation.One (One (One))
 import Pandora.Paradigm.Schemes.T_U (T_U (T_U))