+++ Pandora/Paradigm/Primary/Functor/Convergence.hs
@@ -9,7 +9,7 @@ import Pandora.Pattern.Functor.Monoidal (Monoidal (unit))
 import Pandora.Pattern.Functor (Functor ((-|-)))
 import Pandora.Pattern.Object.Semigroup (Semigroup ((+)))
 import Pandora.Pattern.Object.Monoid (Monoid (zero))
-import Pandora.Paradigm.Algebraic.Exponential (type (-->), type (--<))
+import Pandora.Pattern.Operation.Exponential (type (-->), type (--<))
 import Pandora.Paradigm.Algebraic.Product ((:*:)((:*:)))
 import Pandora.Paradigm.Algebraic ()