+++ Pandora/Paradigm/Primary/Functor/Predicate.hs
@@ -15,7 +15,7 @@ import Pandora.Pattern.Object.Ringoid (Ringoid ((*)))
 import Pandora.Paradigm.Primary.Object.Boolean (Boolean (True, False), bool)
 import Pandora.Paradigm.Algebraic.Product ((:*:)((:*:)))
 import Pandora.Paradigm.Algebraic.Sum ((:+:)(Option, Adoption))
-import Pandora.Paradigm.Algebraic.Exponential (type (-->), type (--<))
+import Pandora.Pattern.Operation.Exponential (type (-->), type (--<))
 
 newtype Predicate a = Predicate (a -> Boolean)