+++ Pandora/Paradigm/Primary/Functor/Wedge.hs
@@ -3,7 +3,7 @@ module Pandora.Paradigm.Primary.Functor.Wedge where
 import Pandora.Pattern.Category ((<--))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<-|-)))
 import Pandora.Pattern.Functor.Traversable (Traversable ((<-/-)))
-import Pandora.Paradigm.Algebraic.Exponential ()
+import Pandora.Pattern.Operation.Exponential ()
 import Pandora.Paradigm.Algebraic (point)
 
 data Wedge e a = Nowhere | Here e | There a