+++ Pandora/Paradigm/Structure/Ability/Monotonic.hs
@@ -2,7 +2,7 @@ module Pandora.Paradigm.Structure.Ability.Monotonic where
 
 import Pandora.Pattern.Category ((<----))
 import Pandora.Pattern.Kernel (constant)
-import Pandora.Paradigm.Algebraic.Exponential ((.:..))
+import Pandora.Pattern.Operation.Exponential ((.:..))
 import Pandora.Paradigm.Algebraic.Sum ((:+:) (Option, Adoption))
 
 class Monotonic a e where