+++ Pandora/Pattern/Operation.hs
@@ -4,3 +4,4 @@ import Pandora.Pattern.Operation.Zero as Exports
 import Pandora.Pattern.Operation.One as Exports
 import Pandora.Pattern.Operation.Sum as Exports
 import Pandora.Pattern.Operation.Product as Exports
+import Pandora.Pattern.Operation.Exponential as Exports